Fix watch (-w)
Delete `nextToken` from kwargs, otherwise this will cause the response
to be stale / the same always.

This should get improved probably to update `kwargs['startTime']`
probably, but it is not clear to me if we could use the `max` or `min`
from the received events here.  Maybe with some offset being applied for
safety (`self.parse_datetime("5m ago")`)?