Configure boto session to use credential cache
This is helpful when using an AWS profile that is assuming a role
with MFA. The user can avoid entering MFA code every time they
execute a command with awslogs.

We specify the same cache directory as is used by AWS CLI, so when
the same profile is used with AWS CLI, the session is shared.