Run JMESPath query when JSON payload is prepended
The --query argument is run only when the message payload starts with
the open curly brace representing a JSON object. This prevented users
from querying JSON data in messages that have prefixed data, e.g. log
level, timestamps, etc.

To work around this, we now locate the first open curly brace (if
present) and attempt to filter as JSON from that point.